reorder only till certain depth, minor bugfix

+++ hash.c
@@ -12,9 +12,9 @@
 /* What, you need more than one hash? Pff. */
 hash_node *hash[HASHSIZE];
 
-unsigned int hash_counter = 0; /* How many slots of the hash are used? */
-unsigned int col_counter  = 0; /* How many collisions happened? */
-unsigned int miss_counter = 0; /* How many entries couldn't be found? */
+unsigned long hash_counter = 0; /* How many slots of the hash are used? */
+unsigned long col_counter  = 0; /* How many collisions happened? */
+unsigned long miss_counter = 0; /* How many entries couldn't be found? */
 
 /* Return result from hash. */
 board_state get_hash(board *board)
@@ -156,8 +156,8 @@ void init_hash()
 /* Prints hash stats. */
 void print_hash_stats()
 {
-	printf("Hash entries: %d, collision: %d, misses: %d, "
-		   "collision percentage: %d%%, used: %d%%.\n",
+	printf("Hash entries: %lu, collision: %lu, misses: %lu, "
+		   "collision percentage: %lu%%, used: %lu%%.\n",
 		   hash_counter, col_counter, miss_counter,
 		   col_counter*100 / (hash_counter > 0 ? hash_counter : 1), 
 		   (hash_counter)*100 / HASHSIZE);
